TEC-6257 Added additional tests for ensuring the JSONMapper is ready for V2 arrays
